# internal/dto/auth.go
package dto

import "github.com/nekoimi/go-project-template/internal/pkg/timeutil"

type RegisterRequest struct {
	Username string `json:"username" binding:"required,min=3,max=50"`
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6,max=50"`
}

type LoginRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required"`
}

type AuthResponse struct {
	Token string   `json:"token"`
	User  UserInfo `json:"user"`
}

type UserInfo struct {
	ID        int64             `json:"id,string"`
	Username  string            `json:"username"`
	Email     string            `json:"email"`
	CreatedAt timeutil.LocalTime `json:"created_at"`
}