# internal/handler/v1/user.go
package v1

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"

	"github.com/nekoimi/go-project-template/internal/pkg/errcode"
	"github.com/nekoimi/go-project-template/internal/pkg/response"
	"github.com/nekoimi/go-project-template/internal/service"
)

type UserHandler struct {
	userService service.UserService
}

func NewUserHandler(userService service.UserService) *UserHandler {
	return &UserHandler{userService: userService}
}

// GetProfile godoc
// @Summary      Get current user profile
// @Tags         users
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  response.APIResponse{data=dto.UserResponse}
// @Failure      401  {object}  response.APIResponse
// @Failure      404  {object}  response.APIResponse
// @Router       /users/profile [get]
func (h *UserHandler) GetProfile(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		response.AppErr(c, errcode.New(errcode.Unauthorized))
		return
	}

	uid, err := strconv.ParseInt(userID.(string), 10, 64)
	if err != nil {
		response.AppErr(c, errcode.New(errcode.Unauthorized))
		return
	}

	profile, err := h.userService.GetProfile(c.Request.Context(), uid)
	if err != nil {
		if appErr, ok := response.IsAppError(err); ok {
			response.AppErr(c, appErr)
			return
		}
		response.ErrorWithMsg(c, http.StatusInternalServerError, errcode.Internal, err.Error())
		return
	}

	response.Success(c, profile)
}